Rewrite stamp test to use bazel_testing
