--- tests/legacy/examples/stamped_bin/stamp/BUILD.bazel
@@ -1,11 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["stamp.go"],
-    importpath = "github.com/bazelbuild/rules_go/examples/stamped_bin/stamp",
-    visibility = ["//visibility:public"],
-    x_defs = {
-        "XdefBuildTimestamp": "{BUILD_TIMESTAMP}",
-    },
-)
