--- tests/legacy/examples/stamped_bin/stamp/stamp.go
@@ -1,28 +1,0 @@
-/* Copyright 2016 The Bazel Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package stamp
-
-var NOT_A_TIMESTAMP = "fail"
-var BUILD_TIMESTAMP = NOT_A_TIMESTAMP
-
-// an xdef should set this to ""
-var PassIfEmpty = "fail"
-
-// an xdef should set this to nonempty
-var XdefBuildTimestamp = ""
-
-// an xdef with a missing key should leave this alone
-var XdefInvalid = "pass"
